Enable execution as sample script from pzs-ng
Add a new target for make "samplescript" that will compile the
binary so it does not execute zipscript-c and removes the rar
parsing. That way you can stick it in your pzs-ng as sample script
if you don't care about the future rar features and don't want to
execute this as a post_check for every file uploaded.